Use test fixture, and move fast tests in their own class
